+++ app/controllers/buyer_details_controller.rb
@@ -1,0 +1,17 @@
+class BuyerDetailsController < ApplicationController
+  def index
+    render status: :not_found
+  end
+
+  def show
+    render status: :not_found
+  end
+
+  def edit
+    render status: :not_found
+  end
+
+  def update
+    render status: :not_found
+  end
+end