+++ app/controllers/legal_panel_for_government/framework_controller.rb
@@ -1,9 +1,10 @@
 module LegalPanelForGovernment
-  class FrameworkController < ::ApplicationController
+  class FrameworkController < ApplicationController
     include FrameworkStatusConcern
 
     before_action :authenticate_user!
     before_action :authorize_user
+    before_action :redirect_to_buyer_detail
 
     protected