+++ app/controllers/legal_services/admin/framework_controller.rb
@@ -1,6 +1,6 @@
 module LegalServices
   module Admin
-    class FrameworkController < ::ApplicationController
+    class FrameworkController < ApplicationController
       include FrameworkStatusConcern
 
       before_action :authenticate_user!