+++ app/controllers/legal_services/framework_controller.rb
@@ -1,5 +1,5 @@
 module LegalServices
-  class FrameworkController < ::ApplicationController
+  class FrameworkController < ApplicationController
     include FrameworkStatusConcern
 
     before_action :authenticate_user!